Develop

+++ src/main/java/co/edu/ufps/innova/consultant/domain/service/ConsultantService.java
@@ -4,56 +4,85 @@ import java.util.List;
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.mail.MailSendException;
+import co.edu.ufps.innova.user.domain.service.IUserService;
 import co.edu.ufps.innova.consultant.domain.dto.Consultant;
-import co.edu.ufps.innova.security.domain.service.UserService;
-import co.edu.ufps.innova.security.domain.service.PasswordService;
+import co.edu.ufps.innova.email.domain.service.IEmailService;
+import co.edu.ufps.innova.authentication.domain.service.IPasswordService;
 import co.edu.ufps.innova.consultant.domain.repository.IConsultantRepository;
 
 @Service
 @RequiredArgsConstructor
-public class ConsultantService {
+public class ConsultantService implements IConsultantService {
 
-    private final UserService userService;
-    private final PasswordService pwService;
+    private final IUserService userService;
+    private final IEmailService emailService;
     private final IConsultantRepository repository;
+    private final IPasswordService passwordService;
 
+    /**
+     * {@inheritDoc}
+     */
+    @Override
     public Consultant save(Consultant consultant) {
-        String userPassword;
-        if (userService.findById(consultant.getId()).isPresent()) {
-            userPassword = userService.getPassword(consultant.getId());
-            userService.delete(consultant.getId());
-        } else {
-            userPassword = pwService.generatePassword();
+        if (userService.findById(consultant.getId()).isPresent()) userService.delete(consultant.getId());
+        String userPassword = passwordService.generatePassword();
+        Consultant newConsultant = repository.save(consultant, passwordService.encryptPassword(userPassword));
+        try {
+            emailService.sendNewUserEmail(newConsultant.getEmail(), userPassword);
+        } catch (MailSendException e) {
+            e.printStackTrace();
         }
-        return repository.save(consultant, userPassword);
+        return newConsultant;
     }
 
+    /**
+     * {@inheritDoc}
+     */
+    @Override
     public boolean update(String id, Consultant consultant) {
         return findById(id)
                 .map(item -> {
+                    consultant.setId(id);
                     repository.save(consultant, userService.getPassword(id));
                     return true;
                 })
                 .orElse(false);
     }
 
-    public Optional<Consultant> findById(String id) {
-        return repository.findById(id);
-    }
-
-    public Optional<Consultant> findByCode(String code) {
-        return repository.findByCode(code);
-    }
-
+    /**
+     * {@inheritDoc}
+     */
+    @Override
     public List<Consultant> findAll() {
         return repository.findAll();
     }
 
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public Optional<Consultant> findById(String id) {
+        return repository.findById(id);
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
     public boolean delete(String id) {
         return findById(id).map(consultant -> {
             repository.delete(consultant);
             return true;
         }).orElse(false);
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public Optional<List<Consultant>> findByActive() {
+        return repository.findByActive();
     }
 
 }
